# Conflicts:
#	modules/core/src/test/java/org/apache/ignite/internal/thread/context/OperationContextAttributesTest.java

+++ modules/core/src/main/java/org/apache/ignite/internal/OperationContexMessage.java
@@ -28,7 +28,7 @@ public class OperationContexMessage implements Message {
 
     /** Bitmask of effective attributes ids. */
     @Order(1)
-    public int idBitmask;
+    public byte idBitmask;
 
     /** Empty constructor for serialization purposes. */
     public OperationContexMessage() {
